requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
